Fix controller device issues

+++ RSDKv4/Input.hpp
@@ -109,8 +109,8 @@ enum ExtraSDLButtons {
     SDL_CONTROLLER_BUTTON_MAX_EXTRA,
 };
 
-void controllerInit(byte controllerID);
-void controllerClose(byte controllerID);
+void controllerInit(int controllerID);
+void controllerClose(int controllerID);
 #endif
 
 #if RETRO_USING_SDL1
